+++ storage/tests/unit/test_blob.py
@@ -23,6 +23,7 @@ import unittest
 
 import google.cloud.storage.blob
 import mock
+import pytest
 import six
 from six.moves import http_client
 
@@ -273,6 +274,14 @@ class Test_Blob(unittest.TestCase):
         bucket = _Bucket()
         blob = self._make_one(blob_name, bucket=bucket)
         self.assertEqual(blob.path, "/b/name/o/Caf%C3%A9")
+
+    def test_bucket_readonly_property(self):
+        blob_name = "BLOB"
+        bucket = _Bucket()
+        other = _Bucket()
+        blob = self._make_one(blob_name, bucket=bucket)
+        with self.assertRaises(AttributeError):
+            blob.bucket = other
 
     def test_client(self):
         blob_name = "BLOB"
@@ -804,7 +813,9 @@ class Test_Blob(unittest.TestCase):
         #       second request.
         headers["range"] = "bytes=3-5"
         headers["accept-encoding"] = "gzip"
-        call = mock.call("GET", expected_url, data=None, headers=headers)
+        call = mock.call(
+            "GET", expected_url, data=None, headers=headers, timeout=mock.ANY
+        )
         self.assertEqual(transport.request.mock_calls, [call, call])
 
     def test__do_download_simple(self):
@@ -832,7 +843,12 @@ class Test_Blob(unittest.TestCase):
         self.assertEqual(file_obj.getvalue(), b"abcdef")
 
         transport.request.assert_called_once_with(
-            "GET", download_url, data=None, headers=headers, stream=True
+            "GET",
+            download_url,
+            data=None,
+            headers=headers,
+            stream=True,
+            timeout=mock.ANY,
         )
 
     def test__do_download_simple_with_range(self):
@@ -861,7 +877,12 @@ class Test_Blob(unittest.TestCase):
         self.assertEqual(headers["range"], "bytes=1-3")
 
         transport.request.assert_called_once_with(
-            "GET", download_url, data=None, headers=headers, stream=True
+            "GET",
+            download_url,
+            data=None,
+            headers=headers,
+            stream=True,
+            timeout=mock.ANY,
         )
 
     def test__do_download_chunked(self):
@@ -887,7 +908,9 @@ class Test_Blob(unittest.TestCase):
         self.assertEqual(transport.request.call_count, 2)
         # ``headers`` was modified (in place) once for each API call.
         self.assertEqual(headers, {"range": "bytes=3-5"})
-        call = mock.call("GET", download_url, data=None, headers=headers)
+        call = mock.call(
+            "GET", download_url, data=None, headers=headers, timeout=mock.ANY
+        )
         self.assertEqual(transport.request.mock_calls, [call, call])
 
     def test__do_download_chunked_with_range(self):
@@ -913,7 +936,9 @@ class Test_Blob(unittest.TestCase):
         self.assertEqual(transport.request.call_count, 2)
         # ``headers`` was modified (in place) once for each API call.
         self.assertEqual(headers, {"range": "bytes=3-4"})
-        call = mock.call("GET", download_url, data=None, headers=headers)
+        call = mock.call(
+            "GET", download_url, data=None, headers=headers, timeout=mock.ANY
+        )
         self.assertEqual(transport.request.mock_calls, [call, call])
 
     def test_download_to_file_with_failure(self):
@@ -947,6 +972,7 @@ class Test_Blob(unittest.TestCase):
             data=None,
             headers={"accept-encoding": "gzip"},
             stream=True,
+            timeout=mock.ANY,
         )
 
     def test_download_to_file_wo_media_link(self):
@@ -1008,6 +1034,7 @@ class Test_Blob(unittest.TestCase):
                 data=None,
                 headers={"accept-encoding": "gzip"},
                 stream=True,
+                timeout=mock.ANY,
             )
 
     def test_download_to_file_default(self):
@@ -1128,6 +1155,7 @@ class Test_Blob(unittest.TestCase):
             data=None,
             headers={"accept-encoding": "gzip"},
             stream=True,
+            timeout=mock.ANY,
         )
 
     def test_download_to_filename_w_key(self):
@@ -1346,7 +1374,7 @@ class Test_Blob(unittest.TestCase):
         )
         headers = {"content-type": b'multipart/related; boundary="==0=="'}
         transport.request.assert_called_once_with(
-            "POST", upload_url, data=payload, headers=headers
+            "POST", upload_url, data=payload, headers=headers, timeout=mock.ANY
         )
 
     @mock.patch(u"google.resumable_media._upload.get_boundary", return_value=b"==0==")
@@ -1505,7 +1533,7 @@ class Test_Blob(unittest.TestCase):
         if extra_headers is not None:
             expected_headers.update(extra_headers)
         transport.request.assert_called_once_with(
-            "POST", upload_url, data=payload, headers=expected_headers
+            "POST", upload_url, data=payload, headers=expected_headers, timeout=mock.ANY
         )
 
     def test__initiate_resumable_upload_no_size(self):
@@ -1579,7 +1607,9 @@ class Test_Blob(unittest.TestCase):
         if size is not None:
             expected_headers["x-upload-content-length"] = str(size)
         payload = json.dumps({"name": blob.name}).encode("utf-8")
-        return mock.call("POST", upload_url, data=payload, headers=expected_headers)
+        return mock.call(
+            "POST", upload_url, data=payload, headers=expected_headers, timeout=mock.ANY
+        )
 
     @staticmethod
     def _do_resumable_upload_call1(
@@ -1596,7 +1626,13 @@ class Test_Blob(unittest.TestCase):
             "content-range": content_range,
         }
         payload = data[: blob.chunk_size]
-        return mock.call("PUT", resumable_url, data=payload, headers=expected_headers)
+        return mock.call(
+            "PUT",
+            resumable_url,
+            data=payload,
+            headers=expected_headers,
+            timeout=mock.ANY,
+        )
 
     @staticmethod
     def _do_resumable_upload_call2(
@@ -1611,7 +1647,13 @@ class Test_Blob(unittest.TestCase):
             "content-range": content_range,
         }
         payload = data[blob.chunk_size :]
-        return mock.call("PUT", resumable_url, data=payload, headers=expected_headers)
+        return mock.call(
+            "PUT",
+            resumable_url,
+            data=payload,
+            headers=expected_headers,
+            timeout=mock.ANY,
+        )
 
     def _do_resumable_helper(
         self, use_size=False, num_retries=None, predefined_acl=None
@@ -1924,7 +1966,7 @@ class Test_Blob(unittest.TestCase):
         if origin is not None:
             expected_headers["Origin"] = origin
         transport.request.assert_called_once_with(
-            "POST", upload_url, data=payload, headers=expected_headers
+            "POST", upload_url, data=payload, headers=expected_headers, timeout=mock.ANY
         )
 
     def test_create_resumable_upload_session(self):
@@ -3129,6 +3171,41 @@ class Test_Blob(unittest.TestCase):
         BUCKET = object()
         blob = self._make_one("blob-name", bucket=BUCKET)
         self.assertIsNone(blob.updated)
+
+    def test_from_string_w_valid_uri(self):
+        from google.cloud.storage.blob import Blob
+
+        connection = _Connection()
+        client = _Client(connection)
+        uri = "gs://BUCKET_NAME/b"
+        blob = Blob.from_string(uri, client)
+
+        self.assertIsInstance(blob, Blob)
+        self.assertIs(blob.client, client)
+        self.assertEqual(blob.name, "b")
+        self.assertEqual(blob.bucket.name, "BUCKET_NAME")
+
+    def test_from_string_w_invalid_uri(self):
+        from google.cloud.storage.blob import Blob
+
+        connection = _Connection()
+        client = _Client(connection)
+
+        with pytest.raises(ValueError, match="URI scheme must be gs"):
+            Blob.from_string("http://bucket_name/b", client)
+
+    def test_from_string_w_domain_name_bucket(self):
+        from google.cloud.storage.blob import Blob
+
+        connection = _Connection()
+        client = _Client(connection)
+        uri = "gs://buckets.example.com/b"
+        blob = Blob.from_string(uri, client)
+
+        self.assertIsInstance(blob, Blob)
+        self.assertIs(blob.client, client)
+        self.assertEqual(blob.name, "b")
+        self.assertEqual(blob.bucket.name, "buckets.example.com")
 
 
 class Test__quote(unittest.TestCase):